MOVE : moves enums from an entity folder to a dto folder
